requirements.txt

torch>=2.1
torchvision>=0.16
timm>=0.9
transformers>=4.41
evaluate>=0.4
scikit-learn>=1.4
numpy>=1.24
pillow>=10.0
tqdm>=4.66
